Adjusted according to review discussion
